Dodaje id w InvoicePositionDto, id Integer -> Long w pozycji faktury

+++ src/main/java/pl/michalkarwowski/api/repositories/InvoicePositionRepository.java
@@ -5,6 +5,6 @@ import pl.michalkarwowski.api.models.InvoicePosition;
 
 import java.util.List;
 
-public interface InvoicePositionRepository extends CrudRepository<InvoicePosition, Integer> {
-    List<InvoicePosition> findAllByProductId(Integer id);
+public interface InvoicePositionRepository extends CrudRepository<InvoicePosition, Long> {
+    List<InvoicePosition> findAllByProductId(Long id);
 }
